Moved CharacterShowScreen to new package
